WIP on speeding up COMPB ISR.  Not working currently.
Instead of looking through the pin array each time through the ISR, we
now pre-compute masks for PORTA..PORTD in the set_phase function.  These
masks, along with the compare values, should make the ISRs take a
fraction of the time which they used to.  It currently looks as if the
math in set_phase which computes the masks is not quite right, though...
COMPA is firing properly, but COMPB never seems to fire.