Multiple changes and bugfixes, but still not quite working
A few changes of note:
-Split off a new drawing of diagrams to help explain the trigonometry
which is required for the various calculations.
-Referenced the above drawings (by figure number) in the source code.
Roughly, figures 1.x are used to understand how to measure the
constants, and figures 2.x are used to understand how the IK
calculations are being obtained.
-Designed a test program which (should) move all legs in synch from the
front of their range of motion to the back (along the y axis), keeping
the x and z values constant.  (It doesn't work currently...)
-Some more work on the simulator (including allowing it to compile under
Linux).