Changing to a 50mm femur, from 20mm.
Changed both the cutting diagram plus the associated measurements in the
code.  Things appear to be working much better with a longer femur;
there is a much larger area which the foot can reach.  The test program
(moving along the Y axis) is not quite perfect, but it is pretty close
to a straight line.
Remaining to do: cut and attach the remaining new femurs, cut new
control rods for tibia + femur, and triple check the measurements (I am
pretty sure that the minor errors in the test program are due to
inaccurate measurements somewhere).