Implemented certain trig functions via lookup tables
Functions include cos, sin, and acos.  Named cos_f (for fast).
Interfaces are in radians, but internally are degrees, using indices
from 0 to 90.  A possible change / enhancement is to convert to a full
255 lookups instead of 90, which would increase the resolution (at the
expense of larger ram usage).

This change increases the code / ram usage by about 2k, but makes the IK
code run about twice as fast.