Change to calibration
We now calibrate against the servo neutral angle (rather than the
desired angle).  The raw value which is stored in EEPROM is multiplied
by 10 and added to the neutral phase when setting the leg positions.  To
find the correct angle to calibrate against, look in Leg.h for the servo
neutral values.