UBBB working.
Some teaks had to be done.  Not sure if this was due to my stupidity or
if this always has to be done.  Tweaks included:
-Flash with LUFA DFU bootloader (changing settings for atmega32u2, 32k
flash, 4k bootloader).
-When programming, hold down PROG and hit RESET.  You should see the
device listed as "Bus 006 Device 048: ID 03eb:2ff0 Atmel Corp." in lsusb
output.  (If it is not in bootloader mode, it does not appear here; I
think this was part of my stupidity in thinking I needed to load the
LUFA bootloader).
-When using the DigitalCave build scripts, set AVRDUDE_PROGRAMMER to
"dfu".  This tells the build scripts to use dfu-programmer (must be
installed) rather than avrdude.