WIP converting Stubby to use new protocol
This is still very much a WIP, and in fact will not work at all at this
moment (although it does compile).  Currently it is expecting messages
in the new protocol format from the universal controller; once the UC
has been updated to send protocol messages, in theory it should work
identically to how it worked before (i.e. the UC can control Stubby
directly, but bluetooth control is not implemented).  However there are
likely additional bugs added during the conversion, so don't even count
on it working as it did before... yet...