Adding preliminary support for magnetometer to Stubby
Basic support is here now... when an API call is made to move in a line,
we check the current heading.  While walking, we then compare the
current heading with the initial heading, and correct as needed.
Initial tests are promising, but there is still a lot of work to do:
-Refactor processing API and protocol to separate move and turn
(programming a veering move would be confusing at best, and it makes the
auto-correcting heading difficult to program)
-Add debug messages to the protocol, and start sending debug info on the
heading correction to verify what is happening
-Possibly look at calibrating the magnetometer on-board stubby (may not
be needed...)