Remove all _delay_ms() calls from Stubby, and replace them with a custom delay function which looks at various mailbox flags and executes non-time sensitive code.  These mailbox flags are set from ISRs.  This allows us to stop running a lot of heavy, yet not time sensitive, code in ISRs and instead run it on the main loop while we normally would be in a delay loop anyway.  Examples of this include the battery level status light, the magnetometer heading, etc.
Another advantage of this is that we now only use a single timer (timer2) to keep track of all of these low-frequency jobs, and all the supporting code is in the same place (hardware/timer2.c).

This code has not yet been tested, but it does compile.  Testing will proceed later tonight if all goes well.