Added a semi-working python pyrawhid sample.
Writing works (and can flash the UBBB's light on and off), but reading
does not work for some reason (the returned value is 0 length).  Writing
is also a bit strange as I can't figure out which byte ends up as the
first byte in the UBBB's buffer.  Still, getting close enough to warrant
a commit.