More changes, currently untested
Some more changes... when wiring up the negative regulator, I found that
the voltage feedback did not work properly, and I needed to buffer the
divided voltage before inverting it.  This required an extra op amp,
which in turn had to be stolen from the instrumentation amp
configuration.  Changes cascade...

This still needs to be tested on the breadboard for both positive and
negative configurations.