Change ArrayStream to call isEmpty() / isFull() instead of duplicating code
I was not sure about this change... with it, the code is simpler but
there is a bit more overhead as we need to do a function call rather
than just a comparison.  As of today I feel that code simplicity is more
important than saving a bit of time... as of tomorrow I will likely feel
differently.
Also added a check in the constructor to verify that capacity is at
least 1 (or else divide by zero errors happen).