Many tweaks:
-Removed Complementary filter completed (Madgwick just works a lot
better)
-Removed SerialUSB support (save some space, we now only support debug
over the air)
-Added a 'mode' parameter to Madgwick compute method.  This will
automatically boost the beta parameter when not armed.  This allows you
to put the battery in while the craft is upside down, and allows it to
quickly orient itself before arming.  TODO we may need to boost this
even more; it is at 100x beta currently, perhaps 200x or more?
Experimentation is needed.
-Changed python script to toggle between rate and negative rate, to show
response times and overshoot.